ui: Add feature to change password
On the user table, there is a button to change the user data. Currently,
only the user's password can be changed. A user can change another
user's password, if the former has the permission 'user.changePassword'
for the latter.
To change another user's password, a user has to enter his/her password
first as a security measure.

E2E tests are added to test for the correct behavior of this feature.